evaluator/requirements.txt
pandas
numpy
anthropic
scikit-learn
openpyxl
ipykernel
python-dotenv
tqdm
fastapi
uvicorn
python-multipart
